+++ crates/idxdb-store/src/ts/chainData.ts
@@ -3,7 +3,7 @@ import { logWebStoreError, uint8ArrayToBase64 } from "./utils.js";
 
 export async function insertBlockHeader(
   dbId: string,
-  blockNum: string,
+  blockNum: number,
   header: Uint8Array,
   partialBlockchainPeaks: Uint8Array,
   hasClientNotes: boolean
@@ -64,7 +64,7 @@ export async function insertPartialBlockchainNodes(
   }
 }
 
-export async function getBlockHeaders(dbId: string, blockNumbers: string[]) {
+export async function getBlockHeaders(dbId: string, blockNumbers: number[]) {
   try {
     const db = getDatabase(dbId);
     const results = await db.blockHeaders.bulkGet(blockNumbers);
@@ -128,7 +128,7 @@ export async function getTrackedBlockHeaders(dbId: string) {
 
 export async function getPartialBlockchainPeaksByBlockNum(
   dbId: string,
-  blockNum: string
+  blockNum: number
 ) {
   try {
     const db = getDatabase(dbId);
@@ -203,7 +203,7 @@ export async function pruneIrrelevantBlocks(dbId: string) {
       .equals("false")
       .and(
         (record) =>
-          record.blockNum !== "0" && record.blockNum !== syncHeight.blockNum
+          record.blockNum !== 0 && record.blockNum !== syncHeight.blockNum
       )
       .toArray();